+++ home/types/react-native-fade-in-image.d.ts
@@ -1,1 +1,0 @@
-declare module 'react-native-fade-in-image';